// src/interfaces/pokemonInterface.tsx
export interface SinglePokemon {
  id: string;
  name: string;
  picture: string;
  color?: string;
}